+++ testdata/xlb/goldens/tsout/localization.ts
@@ -2,10 +2,11 @@
     // Do not modify this file by hand!
     // Re-generate this file by running lit-localize
 
-    import {TemplateResult} from 'lit-html';
-    import {messages as es419Messages} from './es-419.js';
-
     /* eslint-disable @typescript-eslint/no-explicit-any */
+    /* eslint-disable @typescript-eslint/camelcase */
+
+    import {TemplateResult} from 'lit-html';
+    import {messages as es_419Messages} from './es-419.js';
 
     export const supportedLocales = ['en', 'es-419'] as const;
 
@@ -63,7 +64,7 @@
           break;
         
         case 'es-419':
-          resolved = es419Messages[name];
+          resolved = es_419Messages[name];
           break;
         default:
           console.warn(`${locale} is not a supported locale`);